requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.1.18
python-dotenv==1.0.1
django-cors-headers==4.4.0
opencv-python-headless==4.10.0.84
Pillow==10.2.0
